+++ src/helpers/pile-utilities.js
@@ -360,6 +360,8 @@ function getRelevantTokensAndActor(target) {
 
 export async function updateItemPileData(target, flagData, tokenData) {
 
+  Helpers.debug("updateItemPileData on UUID: " + target.uuid);
+
   if (!flagData) flagData = getActorFlagData(target);
   if (!tokenData) tokenData = {};
 
@@ -369,6 +371,8 @@ export async function updateItemPileData(target, flagData, tokenData) {
   const currencies = getActorCurrencies(documentActor, { currencyList: flagData.overrideCurrencies });
 
   const pileData = { data: flagData, items, currencies };
+
+  flagData = foundry.utils.diffObject(CONSTANTS.PILE_DEFAULTS, flagData);
 
   const updates = documentTokens.map(tokenDocument => {
     const newTokenData = foundry.utils.mergeObject(tokenData, {
@@ -397,7 +401,8 @@ export async function updateItemPileData(target, flagData, tokenData) {
 
   if (!foundry.utils.isEmpty(flagData) && documentActor) {
     await documentActor.update({
-      [CONSTANTS.FLAGS.PILE]: flagData, [`token.${CONSTANTS.FLAGS.PILE}`]: flagData
+      [CONSTANTS.FLAGS.PILE]: flagData,
+      [`token.${CONSTANTS.FLAGS.PILE}`]: flagData
     });
   }
 
@@ -405,9 +410,13 @@ export async function updateItemPileData(target, flagData, tokenData) {
 }
 
 export async function updateItemData(item, update) {
-  const flagData = foundry.utils.mergeObject(getItemFlagData(item), update.flags ?? {});
+  const flagData = foundry.utils.diffObject(
+    CONSTANTS.ITEM_DEFAULTS,
+    foundry.utils.mergeObject(getItemFlagData(item), update.flags ?? {})
+  );
   return item.update({
-    ...update?.data ?? {}, [CONSTANTS.FLAGS.ITEM]: flagData
+    ...update?.data ?? {},
+    [CONSTANTS.FLAGS.ITEM]: flagData
   });
 }